+++ packages/badge/src/badge.vue
@@ -1,5 +1,7 @@
 <template>
-  <span :class="dot ? 'weui-badge weui-badge_dot' : 'weui-badge'" :style="style"><slot></slot></span>
+  <span :class="dot ? 'weui-badge_dot' : ''" :style="styleSet" class="weui-badge">
+    <slot></slot>
+  </span>
 </template>
 <script>
 export default {
@@ -9,7 +11,7 @@ export default {
       type: Boolean,
       default: false
     },
-    style: {
+    styleSet: {
       type: String,
       default: null
     }