+++ packages/list/src/group-title.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="weui-cells__title">新消息提示跟摘要信息后，统一在列表右侧</div>
+</template>
+<script>
+export default {
+  name: 'WeuiCellGroupTitle',
+  props: {
+    title: {
+      type: String,
+      default: null
+    }
+  }
+}
+</script>